@@ -1,11 +1,10 @@
 <?php
 
-use Faker\Core\Number;
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class CreateUsersTable extends Migration
+return new class extends Migration
 {
     /**
      * Run the migrations.
@@ -14,24 +13,17 @@ class CreateUsersTable extends Migration
      */
     public function up()
     {
-        Schema::create('users', function (Blueprint $table) { //
+        Schema::create('users', function (Blueprint $table) {
             $table->id();
-            $table->string('first_name');
-            $table->string('last_name');
-            $table->string('username');
-            $table->string('password');
-            $table->date('date_of_birth');
-            $table->string('role_type')->default('user');
-            $table->boolean('license_acceptance')->default(0);
-            $table->string('remember_token')->nullable();
-            $table->timestamps();
-            $table->string('email');
+            $table->string('name');
+            $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
-            $table->string('nationality');
-            $table->integer('confirmation_code')->nullable();
-            $table->timestamp('confirmation_time')->nullable(); //kjndf
+            $table->string('password');
+            $table->rememberToken();
+            $table->timestamps();
         });
     }
+
     /**
      * Reverse the migrations.
      *
@@ -39,7 +31,6 @@ class CreateUsersTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('users'); // kjndf     jsd=jnadfjhbadvklknvfd
-        
+        Schema::dropIfExists('users');
     }
-}
+};
